requirements.txt
websocket-client==1.8.0
numpy==1.26.4
soundfile==0.12.1
tenacity==8.5.0
python-dotenv==1.0.1
rich==13.7.1
httpx==0.27.0